refactor: move headers into namespace folder and add missing include guards
