+++ include/cose/cose.h
@@ -1,5 +1,7 @@
 #include <cn-cbor/cn-cbor.h>
-#include "configure.h"
+#include "cose/configure.h"
+
+#pragma once
 
 #ifdef __cplusplus
 extern "C" {
