+++ src/MacMessage0.c
@@ -10,9 +10,9 @@
 #include <assert.h>
 #include <string.h>
 
-#include "cose.h"
+#include "cose/cose.h"
 #include "cose_int.h"
-#include "configure.h"
+#include "cose/configure.h"
 #include "crypto.h"
 
 #if INCLUDE_MAC0
