+++ src/Sign.c
@@ -4,9 +4,9 @@
 
 #include <stdlib.h>
 
-#include "cose.h"
+#include "cose/cose.h"
 #include "cose_int.h"
-#include "configure.h"
+#include "cose/configure.h"
 
 #if INCLUDE_SIGN
