+++ src/cose_int.h
@@ -1,5 +1,5 @@
-#ifndef __COSE_INT_H__
-#define __COSE_INT_H__
+#pragma once
+
 #include <assert.h>
 
 // These definitions are here because they aren't required for the public
@@ -283,4 +283,3 @@ enum {
 
 
 #define COSE_CounterSign_object 1000
-#endif // __COSE_INT_H__
