+++ test/test.c
@@ -6,8 +6,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <cose.h>
-#include <configure.h>
+#include <cose/cose.h>
+#include <cose/configure.h>
 #include <cn-cbor/cn-cbor.h>
 #include <assert.h>
